+++ lang/en/admin.php
@@ -441,6 +441,7 @@ return [
         'authenticator' => [
             'ldap' => 'LDAP',
             'local' => 'Local',
+            'oidc' => 'OIDC',
             'shibboleth' => 'Shibboleth',
             'title' => 'Authentication Type',
         ],