+++ resources/js/components/MainNav.vue
@@ -362,7 +362,8 @@ async function logout() {
     return;
   }
 
-  await router.push({ name: "logout" });
+  const message = response.data.message || null;
+  await router.push({ name: "logout", query: { message } });
 
   loadingStore.setLoadingFinished();
 }